[Bronze I] Title: 피보나치 수 2, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/2748. 피보나치 수 2/피보나치 수 2.cc
@@ -6,7 +6,7 @@ int main()
 {
     int n = 0;
     cin >> n;
-    vector<long long> fib(n+1);
+    vector<long> fib(n+1);
     
     fib[0] = 0;
     fib[1] = 1;
